requirements.txt
# requirements.txt — Dependências de PRODUÇÃO
# IA Odonto Lab | Sprint 4 — versão consolidada (merge com requirements originais)
#
# Instalar: pip install -r requirements.txt
# Dev:      pip install -r requirements-dev.txt

# ── API ──────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Validação ────────────────────────────────
pydantic==2.10.3

# ── HTTP Client (integração CRM) ─────────────
httpx==0.27.2
requests==2.32.3

# ── Banco de Dados ───────────────────────────
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
pgvector==0.3.6
sqlalchemy==2.0.36

# ── IA / LangChain ───────────────────────────
langchain==0.3.9
langchain-openai==0.2.10
langchain-community==0.3.9
langchain-postgres==0.0.12
langchain-text-splitters==0.3.2
openai==1.56.2

# ── Processamento ────────────────────────────
tiktoken==0.7.0
pypdf==4.3.1

# ── Utilitários ──────────────────────────────
python-dotenv==1.0.1
